- Add Face ID assets - Add BiometricsTypeManager - Add Face ID support to InputTypeViewController & RepearLoginOrRegisterViewController
