+++ LockerUI/LockerUI.xcassets/button-face.imageset/Contents.json
@@ -2,17 +2,17 @@
   "images" : [
     {
       "idiom" : "universal",
-      "filename" : "icons-face-ok.png",
+      "filename" : "button-face.png",
       "scale" : "1x"
     },
     {
       "idiom" : "universal",
-      "filename" : "icons-face-ok@2x.png",
+      "filename" : "button-face@2x.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
-      "filename" : "icons-face-ok@3x.png",
+      "filename" : "button-face@3x.png",
       "scale" : "3x"
     }
   ],
