+++ LockerUI/LockerUI.xcassets/icons-face-broken.imageset/Contents.json
@@ -7,10 +7,12 @@
     },
     {
       "idiom" : "universal",
+      "filename" : "icons-face-broken@2x.png",
       "scale" : "2x"
     },
     {
       "idiom" : "universal",
+      "filename" : "icons-face-broken@3x.png",
       "scale" : "3x"
     }
   ],
